refactor(partners): migrate partner application-intake (+ discord connector) into modules/
