+++ packages/twenty-apps/internal/twenty-partners/src/modules/partner/application-intake/__tests__/submit-partner-application.schema.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 
-import { submitPartnerApplicationSchema } from '../submit-partner-application.logic-function';
+import { submitPartnerApplicationSchema } from '../services/submit-partner-application-input.schema';
 
 const base = {
   firstName: 'Ada',
