+++ packages/twenty-apps/internal/twenty-partners/src/modules/partner/application-intake/connector/discord/types.ts
@@ -1,0 +1,5 @@
+export type DiscordField = { name: string; value: string; inline?: boolean };
+
+export type DiscordEmbed = Record<string, unknown>;
+
+export type DiscordWebhookPayload = { embeds: DiscordEmbed[] };
