+++ packages/twenty-apps/internal/twenty-partners/src/modules/partner/application-intake/graphql/mutations/create-company.ts
@@ -1,0 +1,10 @@
+import type { CoreApiClient, CoreSchema } from 'twenty-client-sdk/core';
+
+export function createCompany(
+  client: CoreApiClient,
+  data: CoreSchema.CompanyCreateInput,
+) {
+  return client.mutation({
+    createCompany: { __args: { data }, id: true },
+  });
+}
