+++ packages/twenty-apps/internal/twenty-partners/src/modules/partner/application-intake/graphql/mutations/create-partner.ts
@@ -1,0 +1,13 @@
+import type { CoreApiClient, CoreSchema } from 'twenty-client-sdk/core';
+
+export function createPartner(
+  client: CoreApiClient,
+  data: CoreSchema.PartnerCreateInput,
+) {
+  return client.mutation({
+    createPartner: {
+      __args: { data },
+      id: true,
+    },
+  });
+}
