+++ packages/twenty-apps/internal/twenty-partners/src/modules/partner/application-intake/graphql/mutations/create-person.ts
@@ -1,0 +1,13 @@
+import type { CoreApiClient, CoreSchema } from 'twenty-client-sdk/core';
+
+export function createPerson(
+  client: CoreApiClient,
+  data: CoreSchema.PersonCreateInput,
+) {
+  return client.mutation({
+    createPerson: {
+      __args: { data },
+      id: true,
+    },
+  });
+}
