+++ packages/twenty-apps/internal/twenty-partners/src/modules/partner/application-intake/graphql/mutations/update-partner.ts
@@ -1,0 +1,14 @@
+import type { CoreApiClient, CoreSchema } from 'twenty-client-sdk/core';
+
+export function updatePartner(
+  client: CoreApiClient,
+  id: string,
+  data: CoreSchema.PartnerUpdateInput,
+) {
+  return client.mutation({
+    updatePartner: {
+      __args: { id, data },
+      id: true,
+    },
+  });
+}
