+++ packages/twenty-apps/internal/twenty-partners/src/modules/partner/application-intake/graphql/mutations/update-person.ts
@@ -1,0 +1,14 @@
+import type { CoreApiClient, CoreSchema } from 'twenty-client-sdk/core';
+
+export function updatePerson(
+  client: CoreApiClient,
+  id: string,
+  data: CoreSchema.PersonUpdateInput,
+) {
+  return client.mutation({
+    updatePerson: {
+      __args: { id, data },
+      id: true,
+    },
+  });
+}
