+++ packages/twenty-apps/internal/twenty-partners/src/modules/partner/application-intake/utils/derive-deployment-expertise.ts
